Updating the start/stop scripts ..

+++ bin/server/stop.sh
@@ -6,4 +6,7 @@ SCRIPT_DIR="$(dirname $(${READLINK} -f $0))"
 
 SERVER_EXEC_NAME="iqlect_time_server"
 
-kill -9 $(${PID_OF} ${SERVER_EXEC_NAME})
+if ps ax | grep -v grep | grep ${SERVER_EXEC_NAME} > /dev/null
+then
+  kill -9 $(${PID_OF} ${SERVER_EXEC_NAME})
+fi
